Feature data overlay - first commit
First commit containing the implementation of feature data overlay
