+++ smile/Deprecatedfns/optical_sensor_legacy.py
